extract getParamsFromSql function
