+++ src/generateParamsTypeFromSql.ts
@@ -1,10 +1,9 @@
 import * as mysql from 'mysql2/promise';
-import { AST, Parser } from 'node-sql-parser';
+
 import { getTableDefinition } from './getTableDefinition';
 import { mapColumnDefinitionToType } from './mapColumnDefinitionToType';
 import { capitalize } from './utils/capitalize';
-
-const OPTIONAL_PARAM_REGEXP = /:[A-Za-z_]+\?/g;
+import { getParamsFromSql } from './getParamsFromSql';
 
 const getParamsTypeName = (queryName: string) =>
   `I${capitalize(queryName)}QueryParams`;
@@ -15,74 +14,33 @@ export const generateParamsTypeFromSql = async (
   queryName: string,
   sql: string,
 ) => {
-  const parser = new Parser();
+  const params = getParamsFromSql(sql);
 
-  // node-sql-parser does not like the syntax we use to define optional params :[A-Za-z_]? and it shouldn't need to support it
-  // so we will pre process the sql string to account for that
-  const optionalParams = (sql.match(OPTIONAL_PARAM_REGEXP) ?? [])
-    // remove starting : and trailing ?
-    .map((v: string) => v.slice(1, -1));
-  const preprocessedSql = optionalParams.length
-    ? // remove trialing ? to not break node-sql-parser
-      sql.replace(OPTIONAL_PARAM_REGEXP, (v) => v.slice(0, -1))
-    : sql;
+  const tables = [...new Set(params.map((p) => p.table))];
 
-  const ast = parser.astify(preprocessedSql) as AST;
+  const mappedParams: string[] = [];
 
-  console.dir(ast, { depth: null });
+  for await (const table of tables) {
+    const tableDefinition = await getTableDefinition(connection, schema, table);
 
-  if (ast.type === 'select' && ast.where) {
-    const paramHolder =
-      ast.where.left.type === 'param'
-        ? 'left'
-        : ast.where.right.type === 'param'
-        ? 'right'
-        : null;
-
-    const columnHolder =
-      ast.where.left.type === 'column_ref'
-        ? 'left'
-        : ast.where.right.type === 'column_ref'
-        ? 'right'
-        : null;
-
-    if (paramHolder && columnHolder) {
-      const param = ast.where[paramHolder].value;
-      const column = ast.where[columnHolder].column;
-      const table =
-        ast.where[columnHolder].table ??
-        (ast.from.length === 1 ? ast.from[0].table : null);
-
-      if (!table) {
-        // MVP: if the query is valid we could go against the db and check to with table the column corresponds
-        // but for now we will ask the user to set aliases when working with multiple tables in a single query
-        throw new Error(
-          `do not know to which table corresponds column ${column}. please specify aliases when working with multiple tables.`,
-        );
-      }
-
-      const tableDefinition = await getTableDefinition(
-        connection,
-        schema,
-        table,
+    for (const param of params.filter((p) => p.table === table)) {
+      const columnDefinition = Object.entries(tableDefinition).find(
+        ([columnName]) => columnName === param.column,
+      )[1];
+      mappedParams.push(
+        `${param.name}${param.optional ? '?' : ''}: ${mapColumnDefinitionToType(
+          columnDefinition,
+        )};`,
       );
-
-      const types = Object.entries(tableDefinition)
-        .filter(([columnName]) => column === columnName)
-        .map(
-          ([, columnDefinition]) =>
-            `${param}${
-              optionalParams.includes(param) ? '?' : ''
-            }: ${mapColumnDefinitionToType(columnDefinition)};`,
-        )
-        .join('\n');
-
-      return `
-        export interface ${getParamsTypeName(queryName)} {
-          ${types}
-        }
-        `;
     }
+  }
+
+  if (mappedParams.length) {
+    return `
+      export interface ${getParamsTypeName(queryName)} {
+        ${mappedParams.join('\n')}
+      }
+      `;
   }
 
   return '';
